Rename a few objects to be more clear
