--- engine/old_engine.go
@@ -1,104 +1,0 @@
-package engine
-
-import (
-	"os"
-	"path/filepath"
-)
-
-// TODO: needed still?
-func NormalizeWorkdir(workdir string) (string, error) {
-	if workdir == "" {
-		workdir = os.Getenv("DAGGER_WORKDIR")
-	}
-
-	if workdir == "" {
-		var err error
-		workdir, err = os.Getwd()
-		if err != nil {
-			return "", err
-		}
-	}
-	workdir, err := filepath.Abs(workdir)
-	if err != nil {
-		return "", err
-	}
-
-	return workdir, nil
-}
-
-/* TODO: some of bits and pieces here still need to be reincorporated
-const (
-	cacheConfigEnvName = "_EXPERIMENTAL_DAGGER_CACHE_CONFIG"
-)
-
-// nolint: gocyclo
-func Start(ctx context.Context, startOpts Config, fn StartCallback) error {
-	// NB(vito): this RootLabels call effectively makes loading labels
-	// synchronous, but it was already required for running just about any query
-	// (see core/query.go), and it's still helpful to have the separate
-	// LoadRootLabels step until we can get rid of the core/query.go call site.
-	labels := []*progrock.Label{}
-	for _, label := range pipeline.RootLabels() {
-		labels = append(labels, &progrock.Label{
-			Name:  label.Name,
-			Value: label.Value,
-		})
-	}
-
-	// Check if any of the upstream cache importers/exporters are enabled.
-	// Note that this is not the cache service support in engine/cache/, that
-	// is a different feature which is configured in the engine daemon.
-	cacheConfigType, cacheConfigAttrs, err := cacheConfigFromEnv()
-	if err != nil {
-		return fmt.Errorf("cache config from env: %w", err)
-	}
-	cacheConfigEnabled := cacheConfigType != ""
-	if cacheConfigEnabled {
-		solveOpts.CacheExports = []bkclient.CacheOptionsEntry{{
-			Type:  cacheConfigType,
-			Attrs: cacheConfigAttrs,
-		}}
-		solveOpts.CacheImports = []bkclient.CacheOptionsEntry{{
-			Type:  cacheConfigType,
-			Attrs: cacheConfigAttrs,
-		}}
-	}
-
-			if cacheConfigEnabled {
-				// Return a result that contains every reference that was solved in this session.
-				return gwClient.CombinedResult(ctx)
-			}
-			return nil, nil
-	})
-
-	return nil
-}
-
-func cacheConfigFromEnv() (string, map[string]string, error) {
-	envVal, ok := os.LookupEnv(cacheConfigEnvName)
-	if !ok {
-		return "", nil, nil
-	}
-
-	// env is in form k1=v1,k2=v2,...
-	kvs := strings.Split(envVal, ",")
-	if len(kvs) == 0 {
-		return "", nil, nil
-	}
-	attrs := make(map[string]string)
-	for _, kv := range kvs {
-		parts := strings.SplitN(kv, "=", 2)
-		if len(parts) != 2 {
-			return "", nil, errors.Errorf("invalid form for cache config %q", kv)
-		}
-		attrs[parts[0]] = parts[1]
-	}
-	typeVal, ok := attrs["type"]
-	if !ok {
-		return "", nil, errors.Errorf("missing type in cache config: %q", envVal)
-	}
-	delete(attrs, "type")
-	return typeVal, attrs, nil
-}
-
-*/
